Reformat code for v1.0 release
