+++ src/main/java/seedu/duke/command/HelpCommand.java
@@ -1,6 +1,5 @@
 package seedu.duke.command;
 
-import seedu.duke.exception.GetJackDException;
 import seedu.duke.lists.WorkoutList;
 import seedu.duke.storage.Storage;
 import seedu.duke.ui.Ui;
@@ -9,8 +8,11 @@ import seedu.duke.ui.Ui;
  * Used to display a help message which shows the user the commands to use.
  */
 public class HelpCommand extends Command {
-    private String commandDescription = null;
     public static final String COMMAND_WORD = "help";
+    private String commandDescription = null;
+
+    public HelpCommand() {
+    }
 
     /**
      * Creates an instance of HelpCommand and sets the help message.
@@ -18,17 +20,22 @@ public class HelpCommand extends Command {
      * use the command.
      * If HelpCommand is constructed without a specified description, then executing the command will just print the
      * generic help message.
+     *
      * @param descriptionToPrint help message to print
      */
     public HelpCommand(String descriptionToPrint) {
         this.commandDescription = descriptionToPrint;
     }
 
-    public HelpCommand() {
-    }
-
+    /**
+     * Executes help command to show user the help message.
+     *
+     * @param workouts List of Workouts
+     * @param ui       User-Interface object
+     * @param storage  Storage object
+     */
     @Override
-    public void executeUserCommand(WorkoutList workouts, Ui ui, Storage storage) throws GetJackDException {
+    public void executeUserCommand(WorkoutList workouts, Ui ui, Storage storage) {
         if (commandDescription == null) {
             Ui.printHelpMessage();
         } else {
