+++ src/main/java/seedu/duke/lists/WorkoutList.java
@@ -7,7 +7,6 @@ import java.util.logging.Logger;
 
 import static seedu.duke.logger.LoggerUtil.setupLogger;
 
-
 /**
  * Represents a list of all workouts in the form of a ArrayList of Workout Objects (All workouts stored by user).
  * Adding, deleting and other operations related to workouts will be executed from here.
@@ -16,9 +15,6 @@ public class WorkoutList {
     private ArrayList<Workout> workouts;
     static final Logger LOGGER = Logger.getLogger(WorkoutList.class.getName());
 
-    /**
-     * Default Constructor.
-     */
     public WorkoutList() {
         workouts = new ArrayList<>();
         setupLogger(LOGGER);
@@ -27,7 +23,7 @@ public class WorkoutList {
     /**
      * Adds a workout routine to the list of workouts.
      *
-     * @param workout workout routine to be added.
+     * @param workout workout routine to be added
      */
     public void addWorkout(Workout workout) {
         assert workout != null;
@@ -38,48 +34,52 @@ public class WorkoutList {
     /**
      * Removes a specific workout routine.
      *
-     * @param displayIndex 1-based index (as seen by user) of the workout routine to be removed.
-     * @return workout routine that was removed.
-     * @throws GetJackDException when the index is greater than number of workouts or less than 0.
+     * @param displayIndex 1-based index (as seen by user) of the workout routine to be removed
+     * @return workout routine that was removed
+     * @throws GetJackDException when the index is greater than number of workouts or less than 0
      */
     public Workout removeWorkout(int displayIndex) throws GetJackDException {
         if (displayIndex <= 0 || displayIndex > workouts.size()) {
             throw new GetJackDException("Invalid Workout Index");
         }
+
         LOGGER.info("Removing Workout");
+
         return workouts.remove(displayIndex - 1);
     }
 
     /**
      * To get all workout routines stored currently.
      *
-     * @return all workout routines stored in the ArrayList.
+     * @return all workout routines stored in the ArrayList
      */
     public ArrayList<Workout> getAllWorkouts() {
         return workouts;
     }
 
     /**
-     * returns a specific workout routine.
+     * Returns a specific workout routine.
      *
-     * @param displayIndex 1-based index (as seen by user) of workout routine to be returned.
-     * @return workout routine that was requested for.
-     * @throws GetJackDException when the index is greater than number of workouts or less than 0.
+     * @param displayIndex 1-based index (as seen by user) of workout routine to be returned
+     * @return workout routine that was requested for
+     * @throws GetJackDException when the index is greater than number of workouts or less than 0
      */
     public Workout getWorkout(int displayIndex) throws GetJackDException {
         if (displayIndex <= 0 || displayIndex > workouts.size()) {
             throw new GetJackDException("Invalid Workout Index");
         }
+
         return workouts.get(displayIndex - 1);
     }
 
     /**
      * Loops through all workout routines and constructs corresponding WorkoutModel objects for each,
      * to be stored in WorkoutListModel class.
-     * This is done as WorkoutModel objects are easily convertable to JSONStrings for storage.
+     * This is done as WorkoutModel objects are easily convertible to JSONStrings for storage.
      */
     public void convertAllWorkoutsToStorageModel() {
         LOGGER.info("Generating WorkoutListModel");
+
         for (Workout workout : workouts) {
             workout.convertToWorkoutStorageModel();
         }
