+++ src/main/java/seedu/duke/storage/Storage.java
@@ -25,11 +25,11 @@ import static seedu.duke.logger.LoggerUtil.setupLogger;
  * To deal with loading tasks from the json file and saving tasks in the json file.
  */
 public class Storage {
+    private static final Logger LOGGER = Logger.getLogger(Storage.class.getName());
     private final String storagePath = "data/workouts.json";
     private final File file = new File(storagePath);
     private final String filePath = file.getAbsolutePath();
-    private WorkoutListModel workoutListModel = new WorkoutListModel();
-    private static final Logger LOGGER = Logger.getLogger(Storage.class.getName());
+    private final WorkoutListModel workoutListModel = new WorkoutListModel();
 
     public Storage() throws GetJackDException {
         setupLogger(LOGGER);
@@ -37,6 +37,7 @@ public class Storage {
             assert file != null;
             file.getParentFile().mkdirs();
             file.createNewFile();
+
             LOGGER.info("Generated Data Folder and storage JSON");
         } catch (IOException e) {
             LOGGER.log(Level.SEVERE, "Data file can't be created. Exception: ", e);
@@ -46,6 +47,7 @@ public class Storage {
 
     /**
      * Loads data into WorkoutList class from JSON file.
+     *
      * @param workoutList Manages workouts after loading data
      * @throws GetJackDException Exception is thrown when data cannot be loaded
      */
@@ -55,9 +57,11 @@ public class Storage {
             return;
         }
         assert file.length() > 0;
+
         try {
             String jsonString = new String(Files.readAllBytes(Paths.get(filePath)));
             ArrayList<WorkoutModel> workoutsStorageForm = convertFromJson(jsonString);
+
             for (WorkoutModel workoutModel : workoutsStorageForm) {
                 Workout workout = new Workout(workoutModel.getWorkoutName());
                 for (ExerciseModel exerciseModel : workoutModel.getExercises()) {
@@ -66,6 +70,7 @@ public class Storage {
                 }
                 workoutList.addWorkout(workout);
             }
+
             LOGGER.info("Successfully loaded data from JSON file.");
         } catch (IOException e) {
             LOGGER.log(Level.SEVERE, "Data file can't be loaded. Exception: ", e);
@@ -75,25 +80,28 @@ public class Storage {
 
     /**
      * Saves data from WorkoutList into the JSON file.
+     *
      * @param jsonString Workout List converted into a JSON string
      * @throws GetJackDException Exception is thrown when there is an error writing to JSON file
      */
     public void saveData(String jsonString) throws GetJackDException {
         assert file.exists();
+
         try {
             FileWriter fileWriter = new FileWriter(storagePath, false);
             fileWriter.write(jsonString);
             fileWriter.close();
+
             LOGGER.info("Successfully saved data into JSON file.");
         } catch (IOException e) {
             LOGGER.log(Level.SEVERE, "Error writing data to save file. Exception: ", e);
             throw new GetJackDException("☹ OOPS!!! Error writing to file while data!");
         }
-
     }
 
     /**
      * Reads exercises as they are read from the JSON file and converts them to Exercise class.
+     *
      * @param exerciseModel Storage model for Exercise class
      * @return Exercise to be stored in the Workout class
      */
@@ -101,14 +109,17 @@ public class Storage {
         int exerciseSets = Integer.parseInt(exerciseModel.getSets());
         int exerciseReps = Integer.parseInt(exerciseModel.getReps());
         Exercise exercise = new Exercise(exerciseModel.getDescription(), exerciseSets, exerciseReps);
+
         if (exerciseModel.getIsDone().equals("true")) {
             exercise.setDone();
         }
+
         return exercise;
     }
 
     /**
      * Converts data read from JSON files and stored data into the attributes of the relevant storage model classes.
+     *
      * @param jsonString jsonString read from JSON file
      * @return Workout List Model
      * @throws GetJackDException Exception is thrown when data from JSON file cannot be converted to storage model class
@@ -119,6 +130,7 @@ public class Storage {
             WorkoutListModel.clearWorkoutListModel();
             assert workoutListModel.getWorkouts().isEmpty();
             WorkoutListModel workoutListModel = JsonUtil.fromJson(node, WorkoutListModel.class);
+
             return workoutListModel.getWorkouts();
         } catch (IOException e) {
             throw new GetJackDException("☹ OOPS!!! Error converting from JSON");
@@ -127,6 +139,7 @@ public class Storage {
 
     /**
      * Converts the workoutList into a JSON string.
+     *
      * @param workoutList Workout List to convert
      * @return Converted JSON string
      * @throws GetJackDException Exception is thrown when workout List cannot be converted
@@ -136,6 +149,7 @@ public class Storage {
         assert workoutListModel.getWorkouts().isEmpty();
         workoutList.convertAllWorkoutsToStorageModel();
         JsonNode node = JsonUtil.toJson(workoutListModel);
+
         try {
             return JsonUtil.stringify(node, true);
         } catch (JsonProcessingException e) {
