+++ src/main/java/seedu/duke/ui/Ui.java
@@ -11,16 +11,10 @@ public class Ui {
     private static final String NEW_LINE = "\n\t";
     private static final String DIVIDER = "________________________________________________________";
 
-    /**
-     * Prints a straight line.
-     */
     public static void printLineSeparator() {
         System.out.println(DIVIDER);
     }
 
-    /**
-     * Prints welcome message.
-     */
     public static void printWelcomeMessage() {
 
         String logo = "\n"
@@ -38,9 +32,6 @@ public class Ui {
         printLineSeparator();
     }
 
-    /**
-     * Prints message to be displayed upon exit of program.
-     */
     public static void printByeMessage() {
         printLineSeparator();
         System.out.println("Bye. Hope you get your desired body soon, have a great day!");
@@ -48,29 +39,26 @@ public class Ui {
 
     }
 
-    /**
-     * Prints help message.
-     */
     public static void printHelpMessage() {
         printLineSeparator();
         System.out.println("Here's a list of commands and what they do.");
         System.out.println("To find out more information about the command, such as input format and parameters, "
                 + "enter \"help COMMAND\" where COMMAND is the command you want to know more about");
-        System.out.println("\"add\" : adds an exercise to a workout");
-        System.out.println("\"done\" : marks an exercise as done");
-        System.out.println("\"remove\" : removes an exercise from a workout");
+        System.out.println("\"add\" : Adds an exercise to a workout");
+        System.out.println("\"done\" : Marks an exercise as done");
+        System.out.println("\"remove\" : Removes an exercise from a workout");
         System.out.println("\"create\" : Creates a new workout");
-        System.out.println("\"delete\" : deletes a workout");
-        System.out.println("\"list\" : lists all your workouts");
-        System.out.println("\"display\" : shows all the exercises in a specified workout");
+        System.out.println("\"delete\" : Deletes a workout");
+        System.out.println("\"list\" : Lists all your workouts");
+        System.out.println("\"display\" : Shows all the exercises in a specified workout");
         System.out.println("\"bye\" : Ends the program");
         printLineSeparator();
     }
 
     /**
-     * Prints error message.
+     * Prints specific error message based on the type of exception.
      *
-     * @param e exception whose message we want to print.
+     * @param e is the exception whose message we want to print
      */
     public static void printErrorMessage(GetJackDException e) {
         printLineSeparator();
@@ -81,7 +69,7 @@ public class Ui {
     /**
      * Prints out all the exercises present in the exercise list of the particular workout.
      *
-     * @param exercises is the list of exercises.
+     * @param exercises is the list of exercises
      */
     public void showExercisesToUser(ArrayList<Exercise> exercises) {
         assert (!exercises.isEmpty());
@@ -91,13 +79,14 @@ public class Ui {
         for (int i = 0; i < exercises.size(); i++) {
             System.out.println((i + 1) + ". " + exercises.get(i));
         }
+
         Ui.printLineSeparator();
     }
 
     /**
      * Prints out all the workouts present in workout list.
      *
-     * @param workouts is the list of workouts.
+     * @param workouts is the list of workouts
      */
     public void showWorkoutsToUser(ArrayList<Workout> workouts) {
         assert (!workouts.isEmpty());
@@ -114,7 +103,7 @@ public class Ui {
     /**
      * Prints the specific message needed at specific points.
      *
-     * @param message is the unique message to be shown.
+     * @param message is the unique message to be shown
      */
     public void showToUser(String message) {
         assert (!message.isEmpty());
